[API] Remove IContainer.getDefaultCharset()
Removes the method IContainer.getDefaultCharset() as not all containers
provide the concept of a default charset.

The method was only used in the FileListFactory to add to the list of
encodings and in the Server file implementation if the file does not
exist.

Furthermore, such a method to access the default charset for a container
should be unnecessary. When creating a file locally, the charset
assigned to the newly created resource should be used instead. When
creating files for a received file activity, the charset contained in
the activity should be used.